+++ src/bullet/index.tsx
@@ -22,7 +22,8 @@ const StyledBullet = styled(Bullet)`
   }
 
   &.kirk-bullet--map-active,
-  &.kirk-bullet--map-inactive {
+  &.kirk-bullet--map-inactive,
+  &.kirk-bullet--search {
     width: ${componentSizes.bulletSizeMap};
     height: ${componentSizes.bulletSizeMap};
     border-width: 3px;
@@ -36,6 +37,11 @@ const StyledBullet = styled(Bullet)`
   &.kirk-bullet--map-inactive {
     border-color: ${color.polylineSecondary};
     box-shadow: 0 0 0 1px ${color.polylineStrokeSecondary};
+  }
+
+  &.kirk-bullet--search {
+    border-width: 4px;
+    border-color: ${color.secondaryText};
   }
 `
 export { BulletTypes, BulletProps } from './Bullet'